backend/requirements.txt
# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# HTTP client (for USDA API)
httpx>=0.25.0

# Rate limiting
slowapi>=0.1.9

# Database
sqlalchemy>=2.0.0

# LLM
openai>=1.3.0

# PDF and image processing
PyMuPDF>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0

# PDF generation
reportlab>=4.0.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
